+++ frontend/package.json
@@ -20,6 +20,7 @@
     "@types/node": "^12.0.0",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
+    "@types/react-hooks-helper": "^1.6.4",
     "@types/react-jsonschema-form": "^1.7.4",
     "@types/react-table": "^7.0.29",
     "axios": "^0.21.1",
@@ -35,6 +36,7 @@
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-google-login": "^5.2.2",
+    "react-hooks-helper": "^1.6.0",
     "react-json-schema": "^1.2.2",
     "react-jsonschema-form": "^1.8.1",
     "react-router-dom": "^5.2.0",
