+++ frontend/src/components/auth/Login.tsx
@@ -7,19 +7,19 @@ import {
   Stack,
   Text,
 } from "@chakra-ui/react";
-import React, { useContext, useState } from "react";
+import React, { useContext } from "react";
+import { useForm } from "react-hooks-helper";
 import { Link, Redirect } from "react-router-dom";
 
 import authAPIClient from "../../APIClients/AuthAPIClient";
-import { HOME_PAGE } from "../../constants/Routes";
+import { HOME_PAGE, SIGNUP_PAGE } from "../../constants/Routes";
 import AuthContext from "../../contexts/AuthContext";
 import { AuthenticatedUser } from "../../types/AuthTypes";
 import Header from "../common/Header";
 
 const Login = (): React.ReactElement => {
   const { authenticatedUser, setAuthenticatedUser } = useContext(AuthContext);
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
+  const [{ email, password }, setValue] = useForm({ email: "", password: "" });
 
   const onLogInClick = async () => {
     const user: AuthenticatedUser = await authAPIClient.login(email, password);
@@ -40,18 +40,15 @@ const Login = (): React.ReactElement => {
         <FormControl mt="2rem">
           <Box>
             <Text textStyle="inputDescription">Email Address</Text>
-            <Input
-              mt="2"
-              value={email}
-              onChange={(event) => setEmail(event.target.value)}
-            />
+            <Input mt="2" value={email} name="email" onChange={setValue} />
           </Box>
           <Box mt="1rem">
             <Text textStyle="inputDescription">Password</Text>
             <Input
               mt="2"
               value={password}
-              onChange={(event) => setPassword(event.target.value)}
+              name="password"
+              onChange={setValue}
             />
           </Box>
           <Text mt="1rem" color="gray.300" textStyle="subtitle">
@@ -81,7 +78,7 @@ const Login = (): React.ReactElement => {
               display="inline-block"
               textDecorationLine="underline"
             >
-              <Link to="/signup"> Sign up now </Link>
+              <Link to={SIGNUP_PAGE}> Sign up now </Link>
             </Text>
           </Stack>
         </Box>
