+++ frontend/src/components/common/icons/index.tsx
@@ -22,20 +22,28 @@ export const CloseIcon = ({ color }: CloseIconProps): JSX.Element => (
 
 export const BackArrow = (): JSX.Element => (
   <Icon viewBox="0 0 24 24">
-    <path
-      d="M11.4375 18.75L4.6875 12L11.4375 5.25"
-      stroke="black"
-      strokeWidth="2.25"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-    <path
-      strokeLinejoin="round"
-      d="M5.625 12H19.3125"
-      stroke="black"
-      strokeWidth="2.25"
-      strokeLinecap="round"
-    />
+    <svg
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M11.4375 18.75L4.6875 12L11.4375 5.25"
+        stroke="black"
+        strokeWidth="2.25"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+      <path
+        strokeLinejoin="round"
+        d="M5.625 12H19.3125"
+        stroke="black"
+        strokeWidth="2.25"
+        strokeLinecap="round"
+      />
+    </svg>
   </Icon>
 );
